src/main/java/uk/gov/digital/ho/pttg/application/namematching/candidates/EntireNonAliasName.java
package uk.gov.digital.ho.pttg.application.namematching.candidates;

import org.springframework.stereotype.Component;
import uk.gov.digital.ho.pttg.application.namematching.CandidateName;
import uk.gov.digital.ho.pttg.application.namematching.InputNames;

import java.util.ArrayList;
import java.util.List;

@Component
public class EntireNonAliasName implements NameMatchingCandidateGenerator {

    @Override
    public List<CandidateName> generateCandidates(InputNames inputNames) {
        List<CandidateName> candidateNames = new ArrayList<>();
        candidateNames.add(new CandidateName(inputNames.fullFirstName(), inputNames.fullLastName()));
        return candidateNames;
    }
}
